vhdl: add renumbering pass
Implement the VHDL renumbering pass. This pass visits the parsed AST,
creates scopes as dictated by the VHDL language reference manual, and
adds definitions for everything that can be referred to by name to these
scopes. Each node that yields a scope or definition is assigned a proper
node ID for later reference.