vhdl: add initial scoreboard setup
Add an initial implementation of a scoreboard. This structure is
intended to provide on-demand compilation throughout moore. Each node is
given a unique ID. To determine e.g. the type or constant value of a
node, the scoreboard is consulted. It will then provide all necessary
intermediate information.

For this to work also add the `typed-arena` crate.