global node ID alloc, improve node storage
Make the allocation of node IDs global in the common crate, by using an
atomic usize. Also improve the NodeStorage and NodeMaker traits by
removing the explicit references and lifetimes, leaving them as a choice
to the implementation. This allows for node storages to accept not only
references, but also tuples of references as values.