add a testing TODO list
Add a `tests/TODO` and a `tests/vhdl/TODO` list which can be used to
capture what has supposedly been implemented in the compiler, but is yet
lacking a proper test. Compiler error messages are a prime example for
this, as they are usually implemented "en passant", but are not
rigorously tested. To perform these tests, a proper machine-readable
output format for diagnostics is needed.