vhdl: emit llhd for empty entities
Extend the VHDL and global scoreboards such that they emit the LLHD code
for empty entities and architectures. This requires some cleaning up in
the VHDL scoreboard and additional information to be carried with the
unwrapped AST in the `ast_table`. All entities are currently assumed to
be empty. To also emit the proper port list, the entity will need to be
lowered to HIR and each declared port will need to have its type
determined.