vhdl: impl range type decls, simple const expr
Implement the declaration of integer types via the `range` construct.
This requires simple evaluation of the constant value of expressions.
There is not really any type checking in place as of now, but that will
follow soon. It is now possible to declare integer types and use them in
an entity's port list. The resulting LLHD will contain appropriate `iN`
types. If the integer has negative range, the type collapses to a void.